Fix name folder
Chờ thống nhất và hoàn thiện sau
