--- lib/screen/welcome.dart
@@ -1,23 +1,0 @@
-import 'package:flutter/foundation.dart'; 
-import 'package:flutter/material.dart';  
-
-class welcum_ extends StatefulWidget {
-  const welcum_({super.key});
-
-  @override
-  State<welcum_> createState() => _welcum_State();
-}
-
-class _welcum_State extends State<welcum_> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Color.fromARGB(255, 79, 187, 90),
-      body: SingleChildScrollView(
-          child: Column(
-            
-          ),
-      ),
-    );
-  }
-}
