+++ lib/start_screen/change_pass.dart
@@ -1,18 +1,17 @@
 
+import 'package:fe_garbage_classification_app/start_screen/signup.dart';
 import 'package:flutter/foundation.dart'; 
-import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart'; 
+import 'package:flutter/material.dart'; 
 import 'package:form_field_validator/form_field_validator.dart';
-import 'package:fe_garbage_classification_app/screen/login.dart';
 
-class Signup_ extends StatefulWidget {
-  const Signup_({super.key});
+class Change_pass_ extends StatefulWidget {
+  const Change_pass_({super.key});
 
   @override
-  State<Signup_> createState() => _Signup_State();
+  State<Change_pass_> createState() => _Change_pass_State();
 }
 
-class _Signup_State extends State<Signup_> {
+class _Change_pass_State extends State<Change_pass_> {
   final _formkey = GlobalKey<FormState>(); 
   bool _isPasswordHidden = true ;
   @override
@@ -20,33 +19,32 @@ class _Signup_State extends State<Signup_> {
     return Scaffold(
         backgroundColor: Colors.white,
          appBar: AppBar(
-          //
           toolbarHeight: 30,
           backgroundColor: Color.fromARGB(0, 61, 233, 4) ,
+          //title: Text('Welcome back !'),
+          //centerTitle: true,
+           //,
          ),
         body: SingleChildScrollView(
+          
           child: Column(
+            
             children: <Widget>[
               // 
-              // Header 
-              Column(
-                children: [
-                  Padding(
-                    padding: const EdgeInsets.all( 0.0),
-                    child: Container(
-                      child: Container(
-                        width: 400,
-                        height: 400,
-                        decoration: BoxDecoration(
-                          borderRadius: BorderRadius.circular(40),
-                          border: Border.all(color: Color(0x4FBB5A)),
-                        ),
-                        child: Image.asset('assets/images/login_img.png'),
-                        ),
-                      ),
-                  
+              // Logo - Image on top login screen 
+              Padding(
+                padding: const EdgeInsets.only( top : 0.0),
+                child: Center(
+                  child: Container(
+                    width: 400,
+                    height: 400,
+                    decoration: BoxDecoration(
+                      borderRadius: BorderRadius.circular(40),
+                      border: Border.all(color: Color(0x4FBB5A)),
+                    ),
+                    child: Image.asset('assets/images/login_img.png'),
+                    ),
                   ),
-                ],
               ),
               Row(
                 children: [
@@ -54,7 +52,7 @@ class _Signup_State extends State<Signup_> {
                     padding: const EdgeInsets.only( left: 25),
                     child: Column(
                       children: <Widget>[
-                        Text(' Get On Board! ', style: TextStyle(fontSize: 25, fontWeight: FontWeight.bold, color: Color.fromARGB(255, 79, 187, 90) ),),
+                        Text(' Do not remember your pass ?  ', style: TextStyle(fontSize: 25, fontWeight: FontWeight.bold,color: Color.fromARGB(255, 79, 187, 90) ),),
                         //Text(' Create your profile and start your journey.'),
                       ],
                     ),
@@ -68,7 +66,7 @@ class _Signup_State extends State<Signup_> {
                     child: Column(
                       children: <Widget>[
                         //Text(' Get On Board! ', style: TextStyle(fontSize: 25, fontWeight: FontWeight.bold ),),
-                        Text(' Create your profile and start your journey.'),
+                        Text(' Dont worry, we can help you create a new one :3 '),
                       ],
                     ),
                   ),
@@ -88,7 +86,6 @@ class _Signup_State extends State<Signup_> {
                         Padding(
                           padding: const EdgeInsets.all(8.0),
                           child: TextFormField(
-                            
                             validator: MultiValidator([
                               RequiredValidator(errorText: 'Enter email address !'),
                               EmailValidator(errorText: 'Correct email filled !'),
@@ -96,11 +93,34 @@ class _Signup_State extends State<Signup_> {
                             decoration: InputDecoration(
                               hintText: 'Enter email' , 
                               labelText:  'Your email' ,
-                              prefixIcon: Icon(Icons.email, color: Color.fromARGB(255, 79, 187, 90),), 
-                              //filled: true ,
-                              //fillColor: Color.fromARGB(0, 108, 79, 187),
+                              prefixIcon: Icon(Icons.email,color: Color.fromARGB(255, 79, 187, 90),), 
+                              suffixIcon: IconButton(
+                                  icon: Icon(Icons.send,color: Color.fromARGB(255, 79, 187, 90),),
+                                  onPressed:(){
+                                    // send sec code to gmail
+                                  },
+                                ),
                               errorStyle: TextStyle(fontSize: 12.0), 
-                             
+                              
+                            ),
+                            
+                          ),
+                        ),
+                        //
+                        // Code box ( about 6 digit) 
+                        Padding(
+                          padding: const EdgeInsets.all(8.0),
+                          child: TextFormField(
+                            validator: MultiValidator([
+                              RequiredValidator(errorText: 'Enter code'),
+                              //EmailValidator(errorText: 'Correct email filled !'),
+                            ]),
+                            decoration: InputDecoration(
+                              hintText: 'Enter code (6 digits)' , 
+                              labelText:  'Security Code' ,
+                              prefixIcon: Icon(Icons.security, color: Color.fromARGB(255, 79, 187, 90),), 
+                              errorStyle: TextStyle(fontSize: 12.0), 
+                              
                             ),
                             
                           ),
@@ -113,23 +133,23 @@ class _Signup_State extends State<Signup_> {
                               validator: MultiValidator([ 
                                 RequiredValidator( 
                                     errorText: 'Please enter Password'), 
-                                MinLengthValidator(6, 
+                                MinLengthValidator(8, 
                                     errorText: 
-                                        'Password must be atlist 6 digit'), 
+                                        'Password must be atlist 8 digit'), 
                                 PatternValidator(r'(?=.*?[#!@$%^&*-])', 
                                     errorText: 
-                                        'Password must be at least one special character') 
+                                        'Password must be atlist one special character') 
                               ]), 
                               obscureText: _isPasswordHidden,
                               decoration: InputDecoration( 
-                                hintText: 'Enter Password', 
-                                labelText: 'Your password', 
+                                hintText: 'Enter new password', 
+                                labelText: 'Your new password', 
                                 prefixIcon: Icon( 
                                   Icons.key, 
                                   color: Color.fromARGB(255, 79, 187, 90), 
                                 ), 
                                 suffixIcon: IconButton(
-                                  icon: Icon(_isPasswordHidden ? Icons.visibility_off: Icons.visibility, color: Color.fromARGB(255, 79, 187, 90), ),
+                                  icon: Icon(_isPasswordHidden ? Icons.visibility_off: Icons.visibility,color: Color.fromARGB(255, 79, 187, 90), ),
                                   onPressed:(){
                                     setState(() {
                                       _isPasswordHidden = !_isPasswordHidden; 
@@ -142,46 +162,8 @@ class _Signup_State extends State<Signup_> {
                             ), 
                         ),
                         //
-                        // Name box
-                        Padding(
-                          padding: const EdgeInsets.all(8.0),
-                          child: TextFormField(
-                            validator: MultiValidator([
-                              RequiredValidator(errorText: 'Enter your name !'),
-                              //EmailValidator(errorText: 'Correct email filled !'),
-                            ]),
-                            decoration: InputDecoration(
-                              hintText: 'Enter Name' , 
-                              labelText:  'Your name' ,
-                              prefixIcon: Icon(Icons.person , color: Color.fromARGB(255, 79, 187, 90), ), 
-                              errorStyle: TextStyle(fontSize: 12.0), 
-                              
-                            ),
-                            
-                          ),
-                        ),
-                        //
-                        // already have account
-                        /*
-                        Row(
-                          children: <Widget>[
-                            Padding( 
-                                padding: EdgeInsets.fromLTRB(0, 0, 0, 0), 
-                                child: TextButton(
-                                  child: Text( 'Already have an account ?', style: TextStyle(color: Colors.black  ),  ),
-                                  onPressed:(){
-                                    try{
-                                      Navigator.push(context, MaterialPageRoute(builder: (context) => Login_()  ));
-                                    }catch(e){
-                                      print(e);
-                                    }
-                                  },
-                                ),
-                            ),
-                            
-                          ],
-                        ),
-                        */
+                        // Forgot password and sign up
+                        
                         //
                         //  Login button 
                         Padding(
@@ -193,7 +175,7 @@ class _Signup_State extends State<Signup_> {
                                 style: ButtonStyle(
                                   backgroundColor: MaterialStateProperty.all(Color.fromARGB(255, 79, 187, 90)),
                                 ),
-                                child: Text( 'Sign Up!', style: TextStyle(color: Colors.white,  fontSize:22 ,  ),) ,                                                              
+                                child: Text( 'Change to new password', style: TextStyle(color: Colors.white,  fontSize:22 ,  ),) ,                                                              
                                 onPressed: (){
                                   if( _formkey.currentState!.validate()){
                                     print('Form submited!'); 
@@ -213,27 +195,34 @@ class _Signup_State extends State<Signup_> {
                             ),
                           ),
                         ),
-                        Row(
+                        Column(
                           mainAxisAlignment: MainAxisAlignment.center,
                           children: [
                             Padding(
                               padding: EdgeInsets.fromLTRB(0, 10, 0, 0),
-                              child: Row(children: [                    
-                                IconButton(
-                                  onPressed:(){}, 
-                                  icon: Image.asset('assets/images/fb_icon.png',
-                                  height: 40,
-                                  width: 40,
+                              child: Column(
+                                //mainAxisAlignment: MainAxisAlignment.start,
+                                children: [                    
+                                
+                                SizedBox(
+                                  width: double.infinity,
+                                  child: OutlinedButton.icon(
+                                    onPressed:(){},
+                                    icon: Image.asset('assets/images/fb_icon.png', width: 20, height: 20 , ),
+                                    label: const Text('Sign in with Facebook. '),
                                   ),
                                 ),
-                                SizedBox(width: 50.0),
-                                IconButton(
-                                  onPressed:(){}, 
-                                  icon: Image.asset('assets/images/gg_icon.png',
-                                  height: 40,
-                                  width: 40,
-                                  )
+                                SizedBox(height: 8.0),
+                                SizedBox(
+                                  width: double.infinity,
+                                  child: OutlinedButton.icon(
+                                    onPressed:(){},
+                                    
+                                    icon: Image.asset('assets/images/gg_icon.png', width: 20, height: 20 , ),
+                                    label: const Text('Sign in with Google. '),
+                                  ),
                                 ),
+
                               ],),
                             ),
                           ],
