+++ lib/start_screen/login.dart
@@ -1,17 +1,20 @@
 
-import 'package:fe_garbage_classification_app/screen/signup.dart';
+import 'dart:ui';
+
+import 'package:fe_garbage_classification_app/start_screen/change_pass.dart';
+import 'package:fe_garbage_classification_app/start_screen/signup.dart';
 import 'package:flutter/foundation.dart'; 
 import 'package:flutter/material.dart'; 
 import 'package:form_field_validator/form_field_validator.dart';
 
-class Change_pass_ extends StatefulWidget {
-  const Change_pass_({super.key});
+class Login_ extends StatefulWidget {
+  const Login_({super.key});
 
   @override
-  State<Change_pass_> createState() => _Change_pass_State();
+  State<Login_> createState() => _Login_State();
 }
 
-class _Change_pass_State extends State<Change_pass_> {
+class _Login_State extends State<Login_> {
   final _formkey = GlobalKey<FormState>(); 
   bool _isPasswordHidden = true ;
   @override
@@ -38,10 +41,6 @@ class _Change_pass_State extends State<Change_pass_> {
                   child: Container(
                     width: 400,
                     height: 400,
-                    decoration: BoxDecoration(
-                      borderRadius: BorderRadius.circular(40),
-                      border: Border.all(color: Color(0x4FBB5A)),
-                    ),
                     child: Image.asset('assets/images/login_img.png'),
                     ),
                   ),
@@ -52,8 +51,7 @@ class _Change_pass_State extends State<Change_pass_> {
                     padding: const EdgeInsets.only( left: 25),
                     child: Column(
                       children: <Widget>[
-                        Text(' Do not remember your pass ?  ', style: TextStyle(fontSize: 25, fontWeight: FontWeight.bold,color: Color.fromARGB(255, 79, 187, 90) ),),
-                        //Text(' Create your profile and start your journey.'),
+                        Text(' Welcome back! ', style: TextStyle(fontSize: 25, fontWeight: FontWeight.bold , color: Color.fromARGB(255, 79, 187, 90)),),
                       ],
                     ),
                   ),
@@ -65,8 +63,7 @@ class _Change_pass_State extends State<Change_pass_> {
                     padding: const EdgeInsets.only( left: 29),
                     child: Column(
                       children: <Widget>[
-                        //Text(' Get On Board! ', style: TextStyle(fontSize: 25, fontWeight: FontWeight.bold ),),
-                        Text(' Dont worry, we can help you create a new one :3 '),
+                        Text(' Sign in to pick up where you left off. '),
                       ],
                     ),
                   ),
@@ -93,34 +90,14 @@ class _Change_pass_State extends State<Change_pass_> {
                             decoration: InputDecoration(
                               hintText: 'Enter email' , 
                               labelText:  'Your email' ,
-                              prefixIcon: Icon(Icons.email,color: Color.fromARGB(255, 79, 187, 90),), 
-                              suffixIcon: IconButton(
-                                  icon: Icon(Icons.send,color: Color.fromARGB(255, 79, 187, 90),),
-                                  onPressed:(){
-                                    // send sec code to gmail
-                                  },
-                                ),
+                              prefixIcon: Icon(Icons.email, color: Color.fromARGB(255, 79, 187, 90),), 
                               errorStyle: TextStyle(fontSize: 12.0), 
                               
-                            ),
-                            
-                          ),
-                        ),
-                        //
-                        // Code box ( about 6 digit) 
-                        Padding(
-                          padding: const EdgeInsets.all(8.0),
-                          child: TextFormField(
-                            validator: MultiValidator([
-                              RequiredValidator(errorText: 'Enter code'),
-                              //EmailValidator(errorText: 'Correct email filled !'),
-                            ]),
-                            decoration: InputDecoration(
-                              hintText: 'Enter code (6 digits)' , 
-                              labelText:  'Security Code' ,
-                              prefixIcon: Icon(Icons.security, color: Color.fromARGB(255, 79, 187, 90),), 
-                              errorStyle: TextStyle(fontSize: 12.0), 
-                              
+                              /*border: OutlineInputBorder(
+                                borderSide: BorderSide(color: Color.fromARGB(255, 79, 187, 90)),
+                                borderRadius: BorderRadius.all(Radius.circular(15.0)), 
+                              ),
+                              */
                             ),
                             
                           ),
@@ -142,14 +119,14 @@ class _Change_pass_State extends State<Change_pass_> {
                               ]), 
                               obscureText: _isPasswordHidden,
                               decoration: InputDecoration( 
-                                hintText: 'Enter new password', 
-                                labelText: 'Your new password', 
+                                hintText: 'Enter Password', 
+                                labelText: 'Your password', 
                                 prefixIcon: Icon( 
                                   Icons.key, 
                                   color: Color.fromARGB(255, 79, 187, 90), 
                                 ), 
                                 suffixIcon: IconButton(
-                                  icon: Icon(_isPasswordHidden ? Icons.visibility_off: Icons.visibility,color: Color.fromARGB(255, 79, 187, 90), ),
+                                  icon: Icon(_isPasswordHidden ? Icons.visibility_off: Icons.visibility, color: Color.fromARGB(255, 79, 187, 90), ),
                                   onPressed:(){
                                     setState(() {
                                       _isPasswordHidden = !_isPasswordHidden; 
@@ -157,13 +134,52 @@ class _Change_pass_State extends State<Change_pass_> {
                                   },
                                 ),
                                 errorStyle: TextStyle(fontSize: 12.0), 
-                                
+                                 
                               ), 
                             ), 
                         ),
                         //
                         // Forgot password and sign up
-                        
+                        Row(
+                          children: <Widget>[
+                            Padding( 
+                                padding: EdgeInsets.fromLTRB(0, 0, 0, 0), 
+                                child: TextButton(
+                                  child: Text( 'Forgot Password', style: TextStyle(color: Color.fromARGB(255, 79, 187, 90) ),  ),
+                                  onPressed:(){
+                                    
+                                    try{
+                                      Navigator.push(context, MaterialPageRoute(builder: (context) => Change_pass_()  ));
+                                    }catch(e){
+                                      print(e);
+                                    }
+                                 
+                                  },
+                                ),
+                            ),
+                            Padding( 
+                                padding: EdgeInsets.fromLTRB(0, 0, 0, 0), 
+                                  child: Text( 'or', style: TextStyle(color: Colors.black  ),  ),                                                                                                                         
+                            ),
+                            Padding( 
+                                padding: EdgeInsets.fromLTRB(0, 0, 0, 0), 
+                                child: TextButton(
+                                  child: Text( 'new user ?', style: TextStyle(color: Color.fromARGB(255, 79, 187, 90)  ),  ),
+                                  onPressed:(){
+                                    
+                                    try{
+                                      Navigator.push(context, MaterialPageRoute(builder: (context) => Signup_()  ));
+                                    }catch(e){
+                                      print(e);
+                                    }
+                                 
+                                  },
+                                ), 
+                                
+                                                       
+                            ),
+                          ],
+                        ),
                         //
                         //  Login button 
                         Padding(
@@ -175,7 +191,7 @@ class _Change_pass_State extends State<Change_pass_> {
                                 style: ButtonStyle(
                                   backgroundColor: MaterialStateProperty.all(Color.fromARGB(255, 79, 187, 90)),
                                 ),
-                                child: Text( 'Change to new password', style: TextStyle(color: Colors.white,  fontSize:22 ,  ),) ,                                                              
+                                child: Text( 'Login', style: TextStyle(color: Colors.white,  fontSize:22 ,  ),) ,                                                              
                                 onPressed: (){
                                   if( _formkey.currentState!.validate()){
                                     print('Form submited!'); 
@@ -184,7 +200,7 @@ class _Change_pass_State extends State<Change_pass_> {
                               ),
                             ),
                             width: MediaQuery.of(context).size.width, 
-                              height: 50, 
+                            height: 50, 
                           ),
                         ),
                         Center(
@@ -195,27 +211,33 @@ class _Change_pass_State extends State<Change_pass_> {
                             ),
                           ),
                         ),
-                        Row(
+                        Column(
                           mainAxisAlignment: MainAxisAlignment.center,
                           children: [
                             Padding(
                               padding: EdgeInsets.fromLTRB(0, 10, 0, 0),
-                              child: Row(children: [                    
-                                IconButton(
-                                  onPressed:(){}, 
-                                  icon: Image.asset('assets/images/fb_icon.png',
-                                  height: 40,
-                                  width: 40,
+                              child: Column(
+                                //mainAxisAlignment: MainAxisAlignment.start,
+                                children: [                    
+                                
+                                SizedBox(
+                                  width: double.infinity,
+                                  child: OutlinedButton.icon(
+                                    onPressed:(){},
+                                    icon: Image.asset('assets/images/fb_icon.png', width: 20, height: 20 , ),
+                                    label: const Text('Sign in with Facebook. '),
                                   ),
                                 ),
-                                SizedBox(width: 50.0),
-                                IconButton(
-                                  onPressed:(){}, 
-                                  icon: Image.asset('assets/images/gg_icon.png',
-                                  height: 40,
-                                  width: 40,
-                                  )
+                                SizedBox(height: 8.0),
+                                SizedBox(
+                                  width: double.infinity,
+                                  child: OutlinedButton.icon(
+                                    onPressed:(){},
+                                    icon: Image.asset('assets/images/gg_icon.png', width: 20, height: 20 , ),
+                                    label: const Text('Sign in with Google. '),
+                                  ),
                                 ),
+
                               ],),
                             ),
                           ],
